requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=9.0.0
numpy==1.23.5
transformers==4.30.2
torch==2.0.1
tensorflow==2.10.0
tensorflow-hub==0.14.0
detoxify==0.5.1